requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
dj-static==0.0.6
Django==4.2
django-crontab==0.7.1
django-stdimage==6.0.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
static3==0.7.0
tzdata==2023.3
urllib3==1.26.15
